refactor: Extract OAuth login workflow to separate oauth.py module
- Created new backend/oauth.py module containing:
  * OAuth provider configuration retrieval
  * Token exchange with OAuth providers (Google, GitHub)
  * User information retrieval from providers
  * User creation and linking logic
  * OAuth account management functions
  * Provider color and display utilities

- Refactored backend/app.py:
  * Removed OAuth helper functions and workflow logic
  * Updated OAuth routes to use oauth.py module functions
  * Simplified route handlers to delegate to oauth module
  * Removed unused requests import
  * Cleaner, more focused application logic

- OAuth routes refactored:
  * /api/v1/auth/oauth/<provider>/authorize - uses handle_oauth_authorize()
  * /api/v1/auth/oauth/<provider>/callback - uses handle_oauth_callback()
  * /api/v1/auth/oauth/<provider>/link - uses handle_oauth_link()
  * /api/v1/auth/oauth/providers - uses get_oauth_providers_list()

- Benefits of refactoring:
  * Better separation of concerns
  * OAuth workflow logic centralized in one module
  * Easier OAuth functionality testing and maintenance
  * Cleaner app.py focused on routes and business logic
  * Better code organization and modularity
  * Easier to extend OAuth functionality

- OAuth administration endpoints remain in app.py:
  * Provider CRUD operations
  * Provider management routes
  * Admin-only OAuth functionality

This refactoring improves code organization while maintaining
all existing OAuth functionality and API endpoints.
